Crate sbr_broadcast

Modules

contagion

message

message_headers

murmur

my_macros

node

sieve

utils

Macros

my_print

Functions

main

send_initialisation_signals

Send signals to initialise the sets which require subscriptions.

setup_node

Setup and initialise a node with given parameters.

trigger_send

Wait for system to finish setup and then send the signal to trigger the Broadcast.