Module sbr_broadcast::contagion

Functions

check_delivery

Check the status of the Delivery replies received. If more than the threshold have been received Probabilistic Reliable Broadcast Deliver the Message.

check_ready

Check the status of the Ready replies received. If more than the threshold have been received, add the Message to the Messages which are Ready. And send it as a Ready Message to the Ready peers.

deliver

Probabilistic Consistent Broadcast Deliver. If the Message is verified, send a Ready of the Message to the Ready peers.

deliver_ready

Deliver a Ready type Message. Check if the sending Node is in the Ready peers and/or Delivery peers, and update the corresponding replies if it is.

init

Initialises the Ready set and Delivery set used in the Contagion algorithm. Sample randomly a number of peers from the system and send them an ReadySubscription.

prb_deliver

Probabilistic Reliable Broadcast Deliver. Save the delivered Message in a file with a unique name. This is used to see which Nodes have delivered which Message.

ready_subscribe

Send ReadySubscription to Ready and Delivery peers.

ready_subscription

Deliver a ReadySubscription type Message. Send all the Messages which are ready to the subscribing Node.