Function sbr_broadcast::contagion::deliver_ready

pub async fn deliver_ready(
    keychain: KeyChain, 
    id: usize, 
    signed_msg: SignedMessage, 
    from: Identity, 
    ready_subscribers: Vec<Identity>, 
    ready_replies: Arc<Mutex<HashMap<Identity, Vec<Message>>>>, 
    duplicate_ready: HashMap<Identity, usize>, 
    delivery_replies: Arc<Mutex<HashMap<Identity, Vec<Message>>>>, 
    duplicate_delivery: HashMap<Identity, usize>, 
    node_sender: Sender<SignedMessage>, 
    ready_messages: Arc<Mutex<Vec<Message>>>, 
    r_thr: usize, 
    d_thr: usize, 
    delivered: Arc<Mutex<Option<Message>>>
)

Deliver a Ready type Message. Check if the sending Node is in the Ready peers and/or Delivery peers, and update the corresponding replies if it is.

Arguments

keychain - KeyChain used to sign the Message.

id - The id of the running Node, used for debug purpose.

signed_msg - The signed Message to deliver.

from - The Identity of the Node sending the Ready.

ready_subscribers - The Ready peers subscribed to this Node.

ready_replies - The Atomic Reference Counter to the Ready replies to update.

duplicate_ready - The HashMap containing information on Ready peers sampled multiple times.

delivery_replies - The Atomic Reference Counter to the Delivery replies to update.

duplicate_delivery - The HashMap containing information on Delivery peers sampled multiple times.

node_sender - The Node’s Sender used to send the Gossip Subscription to the peers.

ready_messages - The Atomic Reference Counter to the Messages which are Ready.

r_thr - The threshold defining if enough Ready replies have been received.

d_thr - The threshold defining if enough Delivery replies have been received.

delivered - The Atomic Reference Counter to the delivered Message.